Extract Messages chat surface
